Доработал swagger, добавил все основные ошибки в примерах, разделил хендлеры для каждого конкретного класса

+++ src/main/java/ru/tsk/springboot/educational_rest_project/dao/EmployeeDAO.java
@@ -6,13 +6,13 @@ import ru.tsk.springboot.educational_rest_project.entity.Employee;
 
 import java.util.List;
 
-public interface EmployeeDAO {
+public interface EmployeeDAO{
 
     List<Employee> getAllEmployees();
 
     void saveEmployee(Employee employee);
 
-    Employee getEmployee(int id);
+    Employee getEmployee(long id);
 
-    void deleteEmployee(int id);
+    void deleteEmployee(long id);
 }
